pyproject.toml
[build-system]
requirements = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama_shim"
version = "0.1.0"
authors = [
    { name="Shane Cantwell", email="shane@vcantwell.com" }
]
description = "A service that bridges LM Studio with tools expecting an Ollama API"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi[all]>=0.95.0",
    "httpx>=0.25.0",
    "python-dotenv",
    "pydantic_settings",
]

[project.scripts]
ollama-shim = "src.main:main"

[project.urls]
Homepage = "https://github.com/shanevcantwell/ollama_shim"
"Bug Tracker" = "https://github.com/shanevcantwell/ollama_shim/issues"

[tool.setuptools.packages.find]
where = ["src"]